feat: add todos with comments hook
